use ptrdiff_t for specpdl counts
* src/cygw32.c:
* src/w32fns.c: Use `ptrdiff_t' for storing specpdl indices.

+++ src/w32fns.c
@@ -7878,7 +7878,7 @@ DEFUN ("x-file-dialog", Fx_file_dialog, Sx_file_dialog, 2, 5, 0,
 #endif	/* !NTGUI_UNICODE */
 
     {
-      int count = SPECPDL_INDEX ();
+      ptrdiff_t count = SPECPDL_INDEX ();
 
       w32_dialog_in_progress (Qt);
